+++ saerojinro-infra/src/main/java/goorm/saerojinro/infra/notification/exception/EmitterNotFoundException.java
@@ -1,4 +1,4 @@
-package goorm.saerojinro.api.notification.presentation.exception;
+package goorm.saerojinro.infra.notification.exception;
 
 import goorm.saerojinro.common.exception.CustomException;
