+++ saerojinro-infra/src/main/java/goorm/saerojinro/infra/notification/sse/NotificationSseSender.java
@@ -1,6 +1,6 @@
 package goorm.saerojinro.infra.notification.sse;
 
-import goorm.saerojinro.infra.notification.sse.response.NotificationSendResponse;
+import goorm.saerojinro.infra.notification.response.NotificationSendResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 import org.springframework.web.servlet.mvc.method.annotation.SseEmitter;
